Mejoro el cambio de los angulos y balas con distan
 - Las mueren despues de una distancia indicada por el arma
 - se verifica mas especificamente para cada caso de colision de balas
   para recalcular los angulos

+++ server/game/gun_entity.h
@@ -17,7 +17,7 @@ protected:
     float x;
     float y;
     uint8_t ammo;
-    uint8_t range;
+    uint16_t range;
     uint8_t bullets_to_shoot;
     uint8_t shooted_bullets;
     int16_t initial_angle;
